rename skypy.halo to skypy.halos
