+++ skypy/__init__.py
@@ -15,6 +15,6 @@ from ._astropy_init import *   # noqa
 __all__ = []
 
 from . import galaxies  # noqa
-from . import halo  # noqa
+from . import halos  # noqa
 from . import pipeline  # noqa
 from . import power_spectrum  # noqa
