--- skypy/halo/tests/test_import.py
@@ -1,5 +1,0 @@
-def test_import():
-    import skypy.halo
-    import skypy.halo.abundance_matching
-    import skypy.halo.mass
-    import skypy.halo.quenching
