+++ skypy/halos/tests/test_import.py
@@ -1,0 +1,5 @@
+def test_import():
+    import skypy.halos
+    import skypy.halos.abundance_matching
+    import skypy.halos.mass
+    import skypy.halos.quenching
